Move wxWebViewIE classes to private header
Cleanup includes.
